requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==37.0.4
decouple==0.0.7
Deprecated==1.2.13
Django==4.1.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
dominate==2.7.0
drf-yasg==1.21.4
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
login==0.0.6
lxml==4.9.1
MarkupSafe==2.0.1
numpy==1.23.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.1
redis==4.3.4
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.20.0
SQLAlchemy==1.4.40
sqlparse==0.2.4
stripe==3.0.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.0.1
wrapt==1.14.1
wsproto==1.1.0
WTForms==3.0.1
